refactor: cart.ts를 관심사별로 유틸로 분리하기
