+++ src/refactoring/utils/priceUtils.ts
@@ -1,0 +1,7 @@
+import { CartItem } from '../../types';
+
+export const calculateItemOriginalTotal = (item: CartItem): number => {
+  const { price } = item.product;
+  const { quantity } = item;
+  return price * quantity;
+};
